seventh commit

+++ src/main/java/com/bookstore/api/bookstore/service/impl/CategoryServiceImpl.java
@@ -3,6 +3,7 @@ package com.bookstore.api.bookstore.service.impl;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.bookstore.api.bookstore.dto.CategoryDto;
@@ -41,8 +42,19 @@ public class CategoryServiceImpl implements CategoryService{
 	}
 
 	@Override
-	public void delete(Integer id) {
-		categoryRepository.deleteById(id);
+	public String delete(Integer id) throws Exception {
+//		Category category = category
+		if (!categoryRepository.existsById(id)) {
+			throw new DataIntegrityViolationException("Id khong ton tai");
+		}
+		
+		try {
+			categoryRepository.deleteById(id);
+			return "Xóa thành công";
+		} catch (Exception e) {
+			return e.getMessage();
+		}
+		
 		
 	}
